+++ server.js
@@ -1,8 +1,7 @@
-var fs = require('fs')
 var path = require('path')
 var express = require('express')
-var webpack = require('webpack')
 var serverRender = require('./dist/server.js')
+var favicon = require('serve-favicon')
 
 var app = express()
 var port = process.env.PORT || 5000
@@ -12,7 +11,7 @@ app.use(express.static(path.join(__dirname, 'dist')))
 
 if (isDev) {
   var config = require('./webpack/webpack.config.dev.client.js')
-  var compiler = webpack(config)
+  var compiler = require('webpack')(config)
   app.use(require('webpack-dev-middleware')(compiler, {
     noInfo: false,
     hot:true,
@@ -23,10 +22,14 @@ if (isDev) {
     }
   }))
   app.use(require('webpack-hot-middleware')(compiler))
+}else{
+  app.use(favicon(path.join(__dirname, 'dist', 'favicon.ico')))
+  app.set('views', path.join(__dirname, 'dist'))
+  app.set('view engine', 'ejs')
 }
 
 app.get('*', function (req, res, next) {
-   serverRender.default(req, res);
+  serverRender.default(req, res);
 })
 
 app.listen(port, function(err) {