+++ src/routes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route, IndexRoute, Redirect } from 'react-router'
-import App from './containers/App'
+import App from './components/App'
 import Home from './components/Home'
 import Article from './components/Article'
 import Login from './components/Login'