+++ src/server.js
@@ -2,22 +2,12 @@ import React from 'react'
 import { renderToString } from 'react-dom/server'
 import { RouterContext, match, createMemoryHistory } from 'react-router'
 import { Provider } from 'react-redux'
-import routes from 'routes'
-import configureStore from 'store/configureStore'
 import reactCookie from 'react-cookie'
 import { fromJS,Map,List } from 'immutable'
-import * as Actions from 'actions'
+import configureStore from './store/configureStore'
+import routes from './routes'
 
 async function fetchAllData(dispatch, components, params) {
-  // const needs = components.reduce( (prev, current) => {
-  //   return (current.need || [])
-  //     .concat((current.WrappedComponent ? current.WrappedComponent.need : []) || [])
-  //     .concat(prev);
-  //   }, []);
-  //   const promises = needs.map(need => {
-  //   	return dispatch(need(params))
-  //   });
-  //   return Promise.all(promises);
   const needs = components
       .filter(x=>x.fetchData)
       .reduce((prev,current)=>{
@@ -26,12 +16,12 @@ async function fetchAllData(dispatch, components, params) {
     	.map(x=>{
         return dispatch(x)
     	})
-  let results = await Promise.all(needs);
-  return results
+  return await Promise.all(needs)
 }
 
 function renderFullPage(renderedContent, initialState) {
-  return `<!doctype html><html>
+  return `<!doctype html>
+  <html>
     <head>
       <base href="/">
       <meta charset="utf-8">
@@ -44,27 +34,27 @@ function renderFullPage(renderedContent, initialState) {
     <body class="day-mode">
     <div class="top-box" id="root">${renderedContent}</div>
     <script>
-      window.__INITIAL_STATE__ = ${JSON.stringify(initialState)};
+      window.__INITIAL_STATE__ = ${JSON.stringify(initialState)}
     </script>
     <script type="text/javascript" charset="utf-8" src="/bundle.js"></script>
-    </body></html>
+    </body>
+  </html>
   `
 }
 export default function render(req, res) {
   reactCookie.plugToRequest(req, res)
 	const history = createMemoryHistory()
-
-
-
+  const token = reactCookie.load('token') || null
 	const store = configureStore({auth:fromJS({
-    token: reactCookie.load('token') || null,
+    token: token,
     user: null
   })}, history)
+
 	match({ routes:routes(), location: req.url }, (error, redirectLocation, renderProps) => {
 	  if (error) {
-	    res.status(500).send(error.message);
+	    res.status(500).send(error.message)
 	  } else if (redirectLocation) {
-	    res.redirect(302, redirectLocation.pathname + redirectLocation.search);
+	    res.redirect(302, redirectLocation.pathname + redirectLocation.search)
 	  } else if (renderProps) {
       return fetchAllData(store.dispatch, renderProps.components, renderProps.params)
   	    .then(html=>{
@@ -74,12 +64,20 @@ export default function render(req, res) {
             </Provider>)
   	    	const componentHTML = renderToString(InitialView)
   	    	const initialState = store.getState()
-  	    	return res.status(200).end(renderFullPage(componentHTML, initialState))
+          if(__DEVSERVER__){
+            return res.status(200).end(renderFullPage(componentHTML, initialState))
+          }else{
+            return res.render('index', {__html__: componentHTML,__state__: JSON.stringify(initialState)})
+          }
   	    }).catch(err => {
-  	    	return res.end(renderFullPage("",{}))
+          if(__DEVSERVER__){
+            return res.end(renderFullPage("",{}))
+          }else{
+            return res.render('index', {__html__: "",__state__: {}})
+          }
   	    })
 	  } else {
-	    res.status(404).send('Not Found');
+	    res.status(404).send('Not Found')
 	  }
 	})
 }