+++ src/utils/authService.js
@@ -14,11 +14,11 @@ export function getCookie(name) {
 }
 
 export function removeCookie(name) {
-  cookie.remove(name, cookieConfig);
+  cookie.remove(name, cookieConfig)
 }
 
 export function signOut() {
-  cookie.remove('token', cookieConfig);
+  cookie.remove('token', cookieConfig)
 }
 
 export function isLogin() {