+++ src/utils/index.js
@@ -1,33 +1,33 @@
 export function customTime(item) {
-	let nowTime = new Date().getTime();
-	let minuteTime = 60*1000;
-	let hourTime = 60*minuteTime;
-	let dayTime = 24*hourTime;
-	let monthTime = dayTime * 30;
-	let yearTime = monthTime * 12;
+	let nowTime = new Date().getTime()
+	let minuteTime = 60*1000
+	let hourTime = 60*minuteTime
+	let dayTime = 24*hourTime
+	let monthTime = dayTime * 30
+	let yearTime = monthTime * 12
 
-  let publishTime = new Date(item).getTime();
-  let historyTime = parseInt(nowTime) - parseInt(publishTime);
-  let descTime;
+  let publishTime = new Date(item).getTime()
+  let historyTime = parseInt(nowTime) - parseInt(publishTime)
+  let descTime
   if(historyTime >= yearTime){
     //按年算
-    descTime = parseInt(historyTime/yearTime) + '年前';
+    descTime = parseInt(historyTime/yearTime) + '年前'
   }else if(historyTime< yearTime && historyTime >= monthTime){
     //按月算
-    descTime = parseInt(historyTime/monthTime) + '月前';
+    descTime = parseInt(historyTime/monthTime) + '月前'
   }else if(historyTime< monthTime && historyTime>= dayTime){
     //按天算
-    descTime = parseInt(historyTime/dayTime) + '天前';
+    descTime = parseInt(historyTime/dayTime) + '天前'
   }else if(historyTime< dayTime && historyTime>= hourTime){
     //按小时算
-    descTime = parseInt(historyTime/hourTime) + '小时前';
+    descTime = parseInt(historyTime/hourTime) + '小时前'
   }else if(historyTime< hourTime && historyTime>= minuteTime){
     //按分钟算
-    descTime = parseInt(historyTime/minuteTime) + '分钟前';
+    descTime = parseInt(historyTime/minuteTime) + '分钟前'
   }else{
-    descTime = '刚刚';
+    descTime = '刚刚'
   }
-  return descTime;
+  return descTime
 }
 
 export function formatDate(time) {
@@ -39,4 +39,3 @@ export function formatDate(time) {
   let minutes = tmpDate.getMinutes()
   return year + '.' + mathon + '.' + day + ' ' + hours + ':' + minutes
 }
-